refactor: 메모 도메인 import 변경 #131
